Move body factor callbacks to Compatibility.Patches and tie it into CE_Utility
