+++ Source/CombatExtended/CombatExtended/CE_Utility.cs
@@ -839,7 +839,12 @@ namespace CombatExtended
                 return new Vector2(1, 1);
             }
 
-            var factors = BoundsInjector.ForPawn(pawn);
+            Vector2 factors;
+            if (Compatibility.Patches.GetCollisionBodyFactors(pawn, out factors)) {
+                return factors;
+            }
+
+            factors = BoundsInjector.ForPawn(pawn);
 
             if (pawn.GetPosture() != PawnPosture.Standing || pawn.Downed)
             {
