+++ Source/CombatExtended/Compatibility/Patches.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System;
 using System.Reflection;
 using System.Text;
+using UnityEngine;
 
 namespace CombatExtended.Compatibility
 {
@@ -11,6 +12,8 @@ namespace CombatExtended.Compatibility
         private List<IPatch> patches;
 
         public static List<Func<IEnumerable<ThingDef>>> UsedAmmoCallbacks = new List<Func<IEnumerable<ThingDef>>>();
+        public static List<Func<Pawn, Tuple<bool, Vector2>>> CollisionBodyFactorCallbacks = new List<Func<Pawn, Tuple<bool, Vector2>>>();
+
         public Patches()
         {
             patches = new List<IPatch>();
@@ -88,6 +91,31 @@ namespace CombatExtended.Compatibility
                     }
                 }
             }
+        }
+
+        private static bool _gcbfactive = false;
+
+        public static void RegisterCollisionBodyFactorCallback(Func<Pawn, Tuple<bool, Vector2>> f) {
+            CollisionBodyFactorCallbacks.Add(f);
+            _gcbfactive = true;
+        }
+
+        public static bool GetCollisionBodyFactors(Pawn pawn, out Vector2 ret)
+        {
+            ret = new Vector2();
+            if (_gcbfactive)
+            {
+                foreach (Func<Pawn, Tuple<bool, Vector2>> f in CollisionBodyFactorCallbacks)
+                {
+                    var r = f(pawn);
+                    if (r.Item1)
+                    {
+                        ret = r.Item2;
+                        return true;
+                    }
+                }
+            }
+            return false;
         }
     }
 }
